Merged slave and master roles into one
The locust_slave and locust_master roles shared almost all
of their code, so I merged them into one role in this commit.
There are now two test scenarios that test the same things
but one tests for master mode and one for slave mode.

To reflect this change, the role was also renamed to, simply,
'locust', instead of 'locust_master' and/or 'locust_slave'.