Moved csv option from common to master/stand-alone
The "csv" option gives the basename for where to store CSV
reports. This used to be a "common" option, suitable for all
modes. (master, slave and stand-alone) However, since the
slave doesn't actually write any CSV reports, the option has
been moved from "common" to "master & stand-alone". This is
basically a documentation change only, since we do not check
which options are used for which mode.