New 'base_dir' option indicates where to put stuff
The new parameter 'base_dir' can be used to decide where the
base directory should be. The default is '/opt/locust.io', which
was the old hardcoded value.

Caveat: on subsequent plays, this parameter will have to be
given again, as the combination of base_dir and instance_name
is used to find the instance, and thus to re-use it's settings.