Skip daemon-reload if installed in state 'stopped'
The handler to trigger a 'systemctl daemon-reload' will now be
skipped if the current play is installing a new instance that
is in state 'stopped'. This way it is possible to install
Locust.io without starting any systemd binaries. This might
be useful, for example when building Locust.io docker containers
that don't have a systemd.